chore: move cli's ci oriented commands in new directory ci
