+++ cli/src/repo/index.js
@@ -1,5 +1,4 @@
-const { run, runGet } = require('../run');
-const { join } = require('path');
+const { run } = require('../run');
 const { goback, goroot } = require('../root');
 
 const dependencies = async () => {
@@ -27,26 +26,8 @@ const install = async () => {
     }
 };
 
-const ciexec = async (module, commands) => {
-    const path = goroot()
-    const modulePath = join(process.cwd(), 'modules', module)
-    try {
-        const diff = await runGet(`git diff $(git merge-base develop HEAD)..HEAD ${modulePath}`)
-        goback(path)
-        if (diff === '') {
-            console.log(`No diff found in ${modulePath}`)
-            process.exit(0)
-        }
-    } catch {
-        goback(path)
-    }
-    await run(commands.join(' '))
-}
-
 module.exports = function(program) {
-
     program
-
         .command('repo_dependencies')
         .description('install global dependencies (mainly clis) for the repo')
         .action(dependencies);
@@ -60,10 +41,5 @@ module.exports = function(program) {
         .command('repo_install')
         .description('install repository dependencies')
         .action(install);
-
-    program
-        .command('ci_exec <module> [commands...]')
-        .description('runs command if code changed on current branch')
-        .action(ciexec)
 }
